refactor(frontend): remove notification GraphQL fallback
