+++ frontend/src/lib/api/notificationPreferences.ts
@@ -1,8 +1,7 @@
-import { Code, ConnectError, createClient } from '@connectrpc/connect';
+import { createClient } from '@connectrpc/connect';
 import { createConnectTransport } from '@connectrpc/connect-web';
 import { NotificationPreferencesService } from '$lib/pb/chatto/api/v1/notification_preferences_connect';
-import { NotificationLevel as ApiNotificationLevel } from '$lib/pb/chatto/api/v1/notification_preferences_pb';
-import { NotificationLevel } from '$lib/gql/graphql';
+import { NotificationLevel } from '$lib/pb/chatto/api/v1/notification_preferences_pb';
 
 export type ConnectAPIConfig = {
   baseUrl: string;
@@ -27,55 +26,14 @@ export async function setRoomNotificationLevel(
   const response = await client.setRoomNotificationLevel(
     {
       roomId,
-      level: notificationLevelToAPI(level)
+      level
     },
     {
       headers: config.bearerToken ? { Authorization: `Bearer ${config.bearerToken}` } : undefined
     }
   );
   return {
-    level: notificationLevelFromAPI(response.level),
-    effectiveLevel: notificationLevelFromAPI(response.effectiveLevel)
+    level: response.level,
+    effectiveLevel: response.effectiveLevel
   };
-}
-
-export function shouldFallbackToGraphQL(err: unknown): boolean {
-  if (!(err instanceof ConnectError)) {
-    return true;
-  }
-  return ![
-    Code.Unauthenticated,
-    Code.PermissionDenied,
-    Code.InvalidArgument,
-    Code.FailedPrecondition
-  ].includes(err.code);
-}
-
-function notificationLevelToAPI(level: NotificationLevel): ApiNotificationLevel {
-  switch (level) {
-    case NotificationLevel.Muted:
-      return ApiNotificationLevel.MUTED;
-    case NotificationLevel.Normal:
-      return ApiNotificationLevel.NORMAL;
-    case NotificationLevel.AllMessages:
-      return ApiNotificationLevel.ALL_MESSAGES;
-    case NotificationLevel.Default:
-    default:
-      return ApiNotificationLevel.DEFAULT;
-  }
-}
-
-function notificationLevelFromAPI(level: ApiNotificationLevel): NotificationLevel {
-  switch (level) {
-    case ApiNotificationLevel.MUTED:
-      return NotificationLevel.Muted;
-    case ApiNotificationLevel.NORMAL:
-      return NotificationLevel.Normal;
-    case ApiNotificationLevel.ALL_MESSAGES:
-      return NotificationLevel.AllMessages;
-    case ApiNotificationLevel.DEFAULT:
-    case ApiNotificationLevel.UNSPECIFIED:
-    default:
-      return NotificationLevel.Default;
-  }
 }
