+++ frontend/src/lib/gql/gql.ts
@@ -57,7 +57,6 @@ type Documents = {
     "\n      mutation MatrixClearUserPerm($input: ClearUserPermissionStateInput!) {\n        clearUserPermissionState(input: $input)\n      }\n    ": typeof types.MatrixClearUserPermDocument,
     "\n            query GetServerNotificationPreferences {\n              server {\n                viewerNotificationPreference {\n                  level\n                  effectiveLevel\n                }\n              }\n              viewer {\n                user {\n                  rooms(type: CHANNEL) {\n                    id\n                    name\n                    viewerNotificationPreference {\n                      level\n                      effectiveLevel\n                    }\n                  }\n                }\n              }\n            }\n          ": typeof types.GetServerNotificationPreferencesDocument,
     "\n            mutation SetServerNotificationLevel($input: SetServerNotificationLevelInput!) {\n              setServerNotificationLevel(input: $input) {\n                level\n                effectiveLevel\n              }\n            }\n          ": typeof types.SetServerNotificationLevelDocument,
-    "\n            mutation SetRoomNotificationLevel($input: SetRoomNotificationLevelInput!) {\n              setRoomNotificationLevel(input: $input) {\n                level\n                effectiveLevel\n              }\n            }\n          ": typeof types.SetRoomNotificationLevelDocument,
     "\n    query UserComboboxMembers($search: String) {\n      server {\n        members(search: $search, limit: 10, offset: 0) {\n          users {\n            id\n            login\n            displayName\n            avatarUrl\n          }\n        }\n      }\n    }\n  ": typeof types.UserComboboxMembersDocument,
     "\n  mutation StartDM($input: StartDMInput!) {\n    startDM(input: $input) {\n      id\n    }\n  }\n": typeof types.StartDmDocument,
     "\n  subscription MyServerEvents {\n    myEvents {\n      id\n      createdAt\n      actorId\n      actor {\n        ...UserAvatarUser\n      }\n      event {\n        __typename\n        # Room payloads — full RoomEventView coverage for the chat surface.\n        ... on MessagePostedEvent {\n          roomId\n          body\n          attachments {\n            ...MessageAttachmentView\n          }\n          linkPreview {\n            ...LinkPreviewView\n          }\n          reactions {\n            emoji\n            count\n            hasReacted\n            users(first: 5) {\n              id\n              displayName\n            }\n          }\n          updatedAt\n          inReplyTo\n          threadRootEventId\n          echoOfEventId\n          echoFromThreadRootEventId\n          channelEchoEventId\n          replyCount\n          lastReplyAt\n          threadParticipants(first: 5) {\n            ...UserAvatarUser\n          }\n          viewerIsFollowingThread\n        }\n        ... on MessageEditedEvent {\n          roomId\n          messageEventId\n          body\n          attachments {\n            ...MessageAttachmentView\n          }\n          linkPreview {\n            ...LinkPreviewView\n          }\n          updatedAt\n        }\n        ... on MessageRetractedEvent {\n          roomId\n          messageEventId\n          retractedReason: reason\n        }\n        ... on UserJoinedRoomEvent {\n          roomId\n        }\n        ... on UserLeftRoomEvent {\n          roomId\n        }\n        ... on RoomCreatedEvent {\n          roomId\n        }\n        ... on RoomUpdatedEvent {\n          roomId\n        }\n        ... on RoomDeletedEvent {\n          roomId\n        }\n        ... on RoomArchivedEvent {\n          roomId\n        }\n        ... on RoomUnarchivedEvent {\n          roomId\n        }\n        ... on RoomUniversalChangedEvent {\n          roomId\n          universal\n        }\n        ... on ReactionAddedEvent {\n          roomId\n          messageEventId\n          emoji\n        }\n        ... on ReactionRemovedEvent {\n          roomId\n          messageEventId\n          emoji\n        }\n        ... on PresenceChangedEvent {\n          status\n        }\n        ... on UserTypingEvent {\n          roomId\n          typingThreadRootEventId: threadRootEventId\n        }\n        ... on AssetProcessingStartedEvent {\n          processingRoomId: roomId\n          assetId\n          processingMessageEventId: messageEventId\n        }\n        ... on AssetProcessingSucceededEvent {\n          processingRoomId: roomId\n          assetId\n          processingMessageEventId: messageEventId\n        }\n        ... on AssetProcessingFailedEvent {\n          processingRoomId: roomId\n          assetId\n          processingMessageEventId: messageEventId\n        }\n        ... on AssetDeletedEvent {\n          deletedRoomId: roomId\n          assetId\n        }\n        ... on ServerMemberDeletedEvent {\n          userId\n        }\n        ... on CallStartedEvent {\n          roomId\n          callId\n        }\n        ... on CallParticipantJoinedEvent {\n          roomId\n          callId\n        }\n        ... on CallParticipantLeftEvent {\n          roomId\n          callId\n        }\n        ... on CallEndedEvent {\n          roomId\n          callId\n        }\n        # Deployment-wide events.\n        ... on ServerUpdatedEvent {\n          name\n          description\n          logoUrl\n          bannerUrl\n        }\n        ... on UserProfileUpdatedEvent {\n          userId\n          displayName\n          avatarUrl\n          login\n        }\n        ... on ServerUserPreferencesUpdatedEvent {\n          timezone\n          timeFormat\n        }\n        ... on NotificationLevelChangedEvent {\n          nlcRoomId: roomId\n          level\n          effectiveLevel\n        }\n        ... on MentionNotificationEvent {\n          roomId\n          room {\n            name\n          }\n          actor {\n            id\n            displayName\n          }\n        }\n        ... on NewDirectMessageNotificationEvent {\n          roomId\n          sender {\n            id\n            displayName\n            avatarUrl\n          }\n          conversationName\n        }\n        ... on NotificationCreatedEvent {\n          notificationId\n          roomId\n          eventId\n          inReplyToId\n        }\n        ... on NotificationDismissedEvent {\n          notificationId\n        }\n        ... on RoomMarkedAsReadEvent {\n          roomId\n        }\n        ... on ThreadFollowChangedEvent {\n          tfcRoomId: roomId\n          tfcThreadRootEventId: threadRootEventId\n          isFollowing\n        }\n        ... on RoomGroupsUpdatedEvent {\n          changed\n        }\n        ... on SessionTerminatedEvent {\n          reason\n        }\n        ... on HeartbeatEvent {\n          alive\n        }\n      }\n    }\n  }\n": typeof types.MyServerEventsDocument,
@@ -215,7 +214,6 @@ const documents: Documents = {
     "\n      mutation MatrixClearUserPerm($input: ClearUserPermissionStateInput!) {\n        clearUserPermissionState(input: $input)\n      }\n    ": types.MatrixClearUserPermDocument,
     "\n            query GetServerNotificationPreferences {\n              server {\n                viewerNotificationPreference {\n                  level\n                  effectiveLevel\n                }\n              }\n              viewer {\n                user {\n                  rooms(type: CHANNEL) {\n                    id\n                    name\n                    viewerNotificationPreference {\n                      level\n                      effectiveLevel\n                    }\n                  }\n                }\n              }\n            }\n          ": types.GetServerNotificationPreferencesDocument,
     "\n            mutation SetServerNotificationLevel($input: SetServerNotificationLevelInput!) {\n              setServerNotificationLevel(input: $input) {\n                level\n                effectiveLevel\n              }\n            }\n          ": types.SetServerNotificationLevelDocument,
-    "\n            mutation SetRoomNotificationLevel($input: SetRoomNotificationLevelInput!) {\n              setRoomNotificationLevel(input: $input) {\n                level\n                effectiveLevel\n              }\n            }\n          ": types.SetRoomNotificationLevelDocument,
     "\n    query UserComboboxMembers($search: String) {\n      server {\n        members(search: $search, limit: 10, offset: 0) {\n          users {\n            id\n            login\n            displayName\n            avatarUrl\n          }\n        }\n      }\n    }\n  ": types.UserComboboxMembersDocument,
     "\n  mutation StartDM($input: StartDMInput!) {\n    startDM(input: $input) {\n      id\n    }\n  }\n": types.StartDmDocument,
     "\n  subscription MyServerEvents {\n    myEvents {\n      id\n      createdAt\n      actorId\n      actor {\n        ...UserAvatarUser\n      }\n      event {\n        __typename\n        # Room payloads — full RoomEventView coverage for the chat surface.\n        ... on MessagePostedEvent {\n          roomId\n          body\n          attachments {\n            ...MessageAttachmentView\n          }\n          linkPreview {\n            ...LinkPreviewView\n          }\n          reactions {\n            emoji\n            count\n            hasReacted\n            users(first: 5) {\n              id\n              displayName\n            }\n          }\n          updatedAt\n          inReplyTo\n          threadRootEventId\n          echoOfEventId\n          echoFromThreadRootEventId\n          channelEchoEventId\n          replyCount\n          lastReplyAt\n          threadParticipants(first: 5) {\n            ...UserAvatarUser\n          }\n          viewerIsFollowingThread\n        }\n        ... on MessageEditedEvent {\n          roomId\n          messageEventId\n          body\n          attachments {\n            ...MessageAttachmentView\n          }\n          linkPreview {\n            ...LinkPreviewView\n          }\n          updatedAt\n        }\n        ... on MessageRetractedEvent {\n          roomId\n          messageEventId\n          retractedReason: reason\n        }\n        ... on UserJoinedRoomEvent {\n          roomId\n        }\n        ... on UserLeftRoomEvent {\n          roomId\n        }\n        ... on RoomCreatedEvent {\n          roomId\n        }\n        ... on RoomUpdatedEvent {\n          roomId\n        }\n        ... on RoomDeletedEvent {\n          roomId\n        }\n        ... on RoomArchivedEvent {\n          roomId\n        }\n        ... on RoomUnarchivedEvent {\n          roomId\n        }\n        ... on RoomUniversalChangedEvent {\n          roomId\n          universal\n        }\n        ... on ReactionAddedEvent {\n          roomId\n          messageEventId\n          emoji\n        }\n        ... on ReactionRemovedEvent {\n          roomId\n          messageEventId\n          emoji\n        }\n        ... on PresenceChangedEvent {\n          status\n        }\n        ... on UserTypingEvent {\n          roomId\n          typingThreadRootEventId: threadRootEventId\n        }\n        ... on AssetProcessingStartedEvent {\n          processingRoomId: roomId\n          assetId\n          processingMessageEventId: messageEventId\n        }\n        ... on AssetProcessingSucceededEvent {\n          processingRoomId: roomId\n          assetId\n          processingMessageEventId: messageEventId\n        }\n        ... on AssetProcessingFailedEvent {\n          processingRoomId: roomId\n          assetId\n          processingMessageEventId: messageEventId\n        }\n        ... on AssetDeletedEvent {\n          deletedRoomId: roomId\n          assetId\n        }\n        ... on ServerMemberDeletedEvent {\n          userId\n        }\n        ... on CallStartedEvent {\n          roomId\n          callId\n        }\n        ... on CallParticipantJoinedEvent {\n          roomId\n          callId\n        }\n        ... on CallParticipantLeftEvent {\n          roomId\n          callId\n        }\n        ... on CallEndedEvent {\n          roomId\n          callId\n        }\n        # Deployment-wide events.\n        ... on ServerUpdatedEvent {\n          name\n          description\n          logoUrl\n          bannerUrl\n        }\n        ... on UserProfileUpdatedEvent {\n          userId\n          displayName\n          avatarUrl\n          login\n        }\n        ... on ServerUserPreferencesUpdatedEvent {\n          timezone\n          timeFormat\n        }\n        ... on NotificationLevelChangedEvent {\n          nlcRoomId: roomId\n          level\n          effectiveLevel\n        }\n        ... on MentionNotificationEvent {\n          roomId\n          room {\n            name\n          }\n          actor {\n            id\n            displayName\n          }\n        }\n        ... on NewDirectMessageNotificationEvent {\n          roomId\n          sender {\n            id\n            displayName\n            avatarUrl\n          }\n          conversationName\n        }\n        ... on NotificationCreatedEvent {\n          notificationId\n          roomId\n          eventId\n          inReplyToId\n        }\n        ... on NotificationDismissedEvent {\n          notificationId\n        }\n        ... on RoomMarkedAsReadEvent {\n          roomId\n        }\n        ... on ThreadFollowChangedEvent {\n          tfcRoomId: roomId\n          tfcThreadRootEventId: threadRootEventId\n          isFollowing\n        }\n        ... on RoomGroupsUpdatedEvent {\n          changed\n        }\n        ... on SessionTerminatedEvent {\n          reason\n        }\n        ... on HeartbeatEvent {\n          alive\n        }\n      }\n    }\n  }\n": types.MyServerEventsDocument,
@@ -516,10 +514,6 @@ export function graphql(source: "\n            query GetServerNotificationPrefer
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
 export function graphql(source: "\n            mutation SetServerNotificationLevel($input: SetServerNotificationLevelInput!) {\n              setServerNotificationLevel(input: $input) {\n                level\n                effectiveLevel\n              }\n            }\n          "): (typeof documents)["\n            mutation SetServerNotificationLevel($input: SetServerNotificationLevelInput!) {\n              setServerNotificationLevel(input: $input) {\n                level\n                effectiveLevel\n              }\n            }\n          "];
-/**
- * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
- */
-export function graphql(source: "\n            mutation SetRoomNotificationLevel($input: SetRoomNotificationLevelInput!) {\n              setRoomNotificationLevel(input: $input) {\n                level\n                effectiveLevel\n              }\n            }\n          "): (typeof documents)["\n            mutation SetRoomNotificationLevel($input: SetRoomNotificationLevelInput!) {\n              setRoomNotificationLevel(input: $input) {\n                level\n                effectiveLevel\n              }\n            }\n          "];
 /**
  * The graphql function is used to parse GraphQL queries into a document that can be used by GraphQL clients.
  */
